feat: move program and stage selection to the input section
