+++ src/components/MainSidebar/InputPanel/InputPanel.js
@@ -1,15 +1,44 @@
+import { useCachedDataQuery } from '@dhis2/analytics'
+import { useDataQuery } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
+import { CenteredContent, CircularLoader, NoticeBox } from '@dhis2/ui'
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
-import { tSetUiInput } from '../../../actions/ui.js'
+import { tSetUiInput, tSetUiProgram } from '../../../actions/ui.js'
+import {
+    PROGRAM_TYPE_WITHOUT_REGISTRATION,
+    PROGRAM_TYPE_WITH_REGISTRATION,
+} from '../../../modules/programTypes.js'
+import { DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY } from '../../../modules/userSettings.js'
 import {
     OUTPUT_TYPE_EVENT,
     OUTPUT_TYPE_ENROLLMENT,
 } from '../../../modules/visualization.js'
-import { sGetUiInput } from '../../../reducers/ui.js'
+import { sGetUiInput, sGetUiProgramId } from '../../../reducers/ui.js'
+import { ProgramSelect } from '../ProgramDimensionsPanel/ProgramSelect.js'
 import { InputOption } from './InputOption.js'
 import styles from './InputPanel.module.css'
+
+const query = {
+    programs: {
+        resource: 'programs',
+        params: ({ nameProp }) => ({
+            fields: [
+                'id',
+                `${nameProp}~rename(name)`,
+                'enrollmentDateLabel',
+                'incidentDateLabel',
+                'programType',
+                'programStages[id,displayName~rename(name),displayExecutionDateLabel,hideDueDate,displayDueDateLabel,repeatable]',
+                'displayIncidentDate',
+                'displayIncidentDateLabel',
+                'displayEnrollmentDateLabel',
+            ],
+            paging: false,
+        }),
+    },
+}
 
 const getLabelForInputType = (type) => {
     switch (type) {
@@ -23,8 +52,51 @@ const getLabelForInputType = (type) => {
 }
 
 const InputPanel = ({ visible }) => {
-    const selectedInput = useSelector(sGetUiInput)
     const dispatch = useDispatch()
+    const selectedInput = useSelector(sGetUiInput)
+    const selectedProgramId = useSelector(sGetUiProgramId)
+    const { currentUser } = useCachedDataQuery()
+    const { fetching, error, data, refetch, called } = useDataQuery(query, {
+        lazy: true,
+    })
+    const filteredPrograms = data?.programs.programs.filter(
+        ({ programType }) =>
+            !(
+                selectedInput.type === OUTPUT_TYPE_ENROLLMENT &&
+                programType === PROGRAM_TYPE_WITHOUT_REGISTRATION
+            )
+    )
+    const selectedProgram =
+        selectedProgramId &&
+        filteredPrograms?.find(({ id }) => id === selectedProgramId)
+    const programType = selectedProgram?.programType
+    const requiredStageSelection =
+        selectedInput.type === OUTPUT_TYPE_EVENT &&
+        programType === PROGRAM_TYPE_WITH_REGISTRATION
+    const setSelectedProgramId = (programId) => {
+        if (programId !== selectedProgramId) {
+            const program = filteredPrograms?.find(({ id }) => id === programId)
+            const stage =
+                // auto-select if a program only has a single stage
+                // and input type is Event
+                selectedInput.type === OUTPUT_TYPE_EVENT &&
+                program?.programStages.length === 1
+                    ? program.programStages[0]
+                    : undefined
+            dispatch(tSetUiProgram({ program, stage }))
+        }
+    }
+
+    useEffect(() => {
+        if (visible && !called) {
+            refetch({
+                nameProp:
+                    currentUser.settings[
+                        DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY
+                    ],
+            })
+        }
+    }, [visible, called, refetch])
 
     if (!visible) {
         return null
@@ -33,6 +105,36 @@ const InputPanel = ({ visible }) => {
     const setSelectedInput = (input) => {
         if (selectedInput.type !== input.type) {
             dispatch(tSetUiInput(input))
+        }
+    }
+
+    const renderProgramSelect = () => {
+        if (error && !fetching) {
+            return (
+                <div className={styles.section}>
+                    <NoticeBox error title={i18n.t('Could not load programs')}>
+                        {error?.message ||
+                            i18n.t(
+                                "The programs couldn't be retrieved. Try again or contact your system administrator."
+                            )}
+                    </NoticeBox>
+                </div>
+            )
+        } else if (fetching) {
+            return (
+                <CenteredContent>
+                    <CircularLoader small />
+                </CenteredContent>
+            )
+        } else if (filteredPrograms) {
+            return (
+                <ProgramSelect
+                    programs={filteredPrograms}
+                    selectedProgram={selectedProgram}
+                    setSelectedProgramId={setSelectedProgramId}
+                    requiredStageSelection={requiredStageSelection}
+                />
+            )
         }
     }
 
@@ -46,7 +148,10 @@ const InputPanel = ({ visible }) => {
                 )}
                 onClick={() => setSelectedInput({ type: OUTPUT_TYPE_EVENT })}
                 selected={selectedInput.type === OUTPUT_TYPE_EVENT}
-            />
+            >
+                {selectedInput.type === OUTPUT_TYPE_EVENT &&
+                    renderProgramSelect()}
+            </InputOption>
             <InputOption
                 dataTest="input-enrollment"
                 header={getLabelForInputType(OUTPUT_TYPE_ENROLLMENT)}
@@ -57,7 +162,10 @@ const InputPanel = ({ visible }) => {
                     setSelectedInput({ type: OUTPUT_TYPE_ENROLLMENT })
                 }
                 selected={selectedInput.type === OUTPUT_TYPE_ENROLLMENT}
-            />
+            >
+                {selectedInput.type === OUTPUT_TYPE_ENROLLMENT &&
+                    renderProgramSelect()}
+            </InputOption>
         </div>
     )
 }
