+++ src/components/MainSidebar/ProgramDimensionsPanel/ProgramDimensionsFilter.module.css
@@ -37,3 +37,11 @@
 .option.disabled:hover {
     background-color: initial;
 }
+
+.header {
+    color: var(--colors-grey600);
+    font-size: 11px;
+    letter-spacing: 0.3px;
+    text-transform: uppercase;
+    margin-bottom: var(--spacers-dp8);
+}
