+++ src/components/MainSidebar/ProgramDimensionsPanel/ProgramDimensionsPanel.js
@@ -1,26 +1,17 @@
 import {
-    useCachedDataQuery,
     DIMENSION_TYPE_ALL,
     DIMENSION_TYPE_DATA_ELEMENT,
 } from '@dhis2/analytics'
-import { useDataQuery } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
-import { NoticeBox, CenteredContent, CircularLoader } from '@dhis2/ui'
-import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { tSetUiProgram } from '../../../actions/ui.js'
-import {
-    PROGRAM_TYPE_WITH_REGISTRATION,
-    PROGRAM_TYPE_WITHOUT_REGISTRATION,
-} from '../../../modules/programTypes.js'
-import { DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY } from '../../../modules/userSettings.js'
+import { useSelector } from 'react-redux'
 import { useDebounce } from '../../../modules/utils.js'
 import {
     OUTPUT_TYPE_EVENT,
     OUTPUT_TYPE_ENROLLMENT,
 } from '../../../modules/visualization.js'
+import { sGetMetadataById } from '../../../reducers/metadata.js'
 import {
     sGetUiInputType,
     sGetUiProgramId,
@@ -30,83 +21,23 @@ import { ProgramDataDimensionsList } from './ProgramDataDimensionsList.js'
 import { ProgramDimensions } from './ProgramDimensions.js'
 import { ProgramDimensionsFilter } from './ProgramDimensionsFilter.js'
 import styles from './ProgramDimensionsPanel.module.css'
-import { ProgramSelect } from './ProgramSelect.js'
-
-const query = {
-    programs: {
-        resource: 'programs',
-        params: ({ nameProp }) => ({
-            fields: [
-                'id',
-                `${nameProp}~rename(name)`,
-                'enrollmentDateLabel',
-                'incidentDateLabel',
-                'programType',
-                'programStages[id,displayName~rename(name),displayExecutionDateLabel,hideDueDate,displayDueDateLabel,repeatable]',
-                'displayIncidentDate',
-                'displayIncidentDateLabel',
-                'displayEnrollmentDateLabel',
-            ],
-            paging: false,
-        }),
-    },
-}
 
 const ProgramDimensionsPanel = ({ visible }) => {
-    const dispatch = useDispatch()
     const inputType = useSelector(sGetUiInputType)
     const selectedProgramId = useSelector(sGetUiProgramId)
+    const selectedProgram = useSelector((state) =>
+        sGetMetadataById(state, selectedProgramId)
+    )
     const selectedStageId = useSelector(sGetUiProgramStageId)
-    const { currentUser } = useCachedDataQuery()
-    const { fetching, error, data, refetch, called } = useDataQuery(query, {
-        lazy: true,
-    })
+
     const [searchTerm, setSearchTerm] = useState('')
     const [stageFilter, setStageFilter] = useState()
     const [dimensionType, setDimensionType] = useState(DIMENSION_TYPE_ALL)
     const debouncedSearchTerm = useDebounce(searchTerm)
-    const filteredPrograms = data?.programs.programs.filter(
-        ({ programType }) =>
-            !(
-                inputType === OUTPUT_TYPE_ENROLLMENT &&
-                programType === PROGRAM_TYPE_WITHOUT_REGISTRATION
-            )
-    )
-    const selectedProgram =
-        selectedProgramId &&
-        filteredPrograms?.find(({ id }) => id === selectedProgramId)
-    const programType = selectedProgram?.programType
-    const requiredStageSelection =
-        inputType === OUTPUT_TYPE_EVENT &&
-        programType === PROGRAM_TYPE_WITH_REGISTRATION
     const isProgramSelectionComplete =
         inputType === OUTPUT_TYPE_EVENT
             ? selectedProgram && selectedStageId
             : !!selectedProgram
-    const setSelectedProgramId = (programId) => {
-        if (programId !== selectedProgramId) {
-            const program = filteredPrograms?.find(({ id }) => id === programId)
-            const stage =
-                // auto-select if a program only has a single stage
-                // and input type is Event
-                inputType === OUTPUT_TYPE_EVENT &&
-                program?.programStages.length === 1
-                    ? program.programStages[0]
-                    : undefined
-            dispatch(tSetUiProgram({ program, stage }))
-        }
-    }
-
-    useEffect(() => {
-        if (visible && !called) {
-            refetch({
-                nameProp:
-                    currentUser.settings[
-                        DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY
-                    ],
-            })
-        }
-    }, [visible, called, refetch])
 
     useEffect(() => {
         setSearchTerm('')
@@ -114,87 +45,50 @@ const ProgramDimensionsPanel = ({ visible }) => {
         setStageFilter()
     }, [inputType, selectedProgramId, selectedStageId])
 
-    if (!visible || !called) {
+    if (!visible) {
         return null
-    }
-
-    if (error && !fetching) {
-        return (
-            <div className={styles.section}>
-                <NoticeBox error title={i18n.t('Could not load programs')}>
-                    {error?.message ||
-                        i18n.t(
-                            "The programs couldn't be retrieved. Try again or contact your system administrator."
-                        )}
-                </NoticeBox>
-            </div>
-        )
-    }
-
-    if (fetching) {
-        return (
-            <CenteredContent>
-                <CircularLoader small />
-            </CenteredContent>
-        )
     }
 
     return (
         <div className={styles.container}>
-            <div className={cx(styles.section, styles.bordered)}>
-                <ProgramSelect
-                    programs={filteredPrograms}
-                    selectedProgram={selectedProgram}
-                    setSelectedProgramId={setSelectedProgramId}
-                    requiredStageSelection={requiredStageSelection}
-                />
-            </div>
-            <div>
-                <ProgramDimensions />
-            </div>
-            <div
-                className={cx(styles.section, {
-                    [styles.bordered]: !!selectedProgramId,
-                })}
-            >
-                {isProgramSelectionComplete ? (
-                    <ProgramDimensionsFilter
-                        program={selectedProgram}
-                        searchTerm={searchTerm}
-                        setSearchTerm={setSearchTerm}
-                        dimensionType={dimensionType}
-                        setDimensionType={setDimensionType}
-                        stageFilter={stageFilter}
-                        setStageFilter={setStageFilter}
-                    />
-                ) : (
-                    <div className={styles.helptext}>
-                        {requiredStageSelection
-                            ? i18n.t(
-                                  'Choose a stage above to add program dimensions.'
-                              )
-                            : i18n.t(
-                                  'Choose a program above to add program dimensions.'
-                              )}
+            {isProgramSelectionComplete ? (
+                <>
+                    <div>
+                        <ProgramDimensions />
                     </div>
-                )}
-            </div>
+                    <div className={styles.section}>
+                        <ProgramDimensionsFilter
+                            program={selectedProgram}
+                            searchTerm={searchTerm}
+                            setSearchTerm={setSearchTerm}
+                            dimensionType={dimensionType}
+                            setDimensionType={setDimensionType}
+                            stageFilter={stageFilter}
+                            setStageFilter={setStageFilter}
+                        />
+                    </div>
 
-            {isProgramSelectionComplete && (
-                <ProgramDataDimensionsList
-                    inputType={inputType}
-                    program={selectedProgram}
-                    dimensionType={dimensionType}
-                    searchTerm={debouncedSearchTerm}
-                    stageId={
-                        inputType === OUTPUT_TYPE_ENROLLMENT &&
-                        dimensionType === DIMENSION_TYPE_DATA_ELEMENT
-                            ? stageFilter
-                            : inputType === OUTPUT_TYPE_EVENT
-                            ? selectedStageId
-                            : undefined
-                    }
-                />
+                    <ProgramDataDimensionsList
+                        inputType={inputType}
+                        program={selectedProgram}
+                        dimensionType={dimensionType}
+                        searchTerm={debouncedSearchTerm}
+                        stageId={
+                            inputType === OUTPUT_TYPE_ENROLLMENT &&
+                            dimensionType === DIMENSION_TYPE_DATA_ELEMENT
+                                ? stageFilter
+                                : inputType === OUTPUT_TYPE_EVENT
+                                ? selectedStageId
+                                : undefined
+                        }
+                    />
+                </>
+            ) : (
+                <div className={styles.helptext}>
+                    {i18n.t(
+                        'Choose an input to get started adding program dimensions.'
+                    )}
+                </div>
             )}
         </div>
     )
