+++ src/components/MainSidebar/ProgramDimensionsPanel/ProgramDimensionsPanel.module.css
@@ -6,9 +6,6 @@
 .section {
     padding: var(--spacers-dp12) var(--spacers-dp8);
 }
-.bordered {
-    border-bottom: 1px solid var(--colors-grey300);
-}
 .helptext {
     padding: var(--spacers-dp16) var(--spacers-dp24) 0;
     color: var(--colors-grey600);
