+++ src/components/MainSidebar/ProgramDimensionsPanel/ProgramSelect.js
@@ -1,5 +1,5 @@
 import i18n from '@dhis2/d2-i18n'
-import { SingleSelect, SingleSelectOption, Button, Tooltip } from '@dhis2/ui'
+import { SingleSelect, SingleSelectOption } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React from 'react'
 import styles from './ProgramSelect.module.css'
@@ -11,55 +11,32 @@ const ProgramSelect = ({
     requiredStageSelection,
     selectedProgram,
 }) => {
-    /*
-     * TODO: the logic for disabling the select and showing the clear
-     * button needs to be changed later on. Currently this happens when
-     * programId is present but it should actually happen once at least one
-     * dimension item is selected
-     */
     const showStageSelect = selectedProgram && requiredStageSelection
-    const select = (
-        <SingleSelect
-            dense
-            selected={selectedProgram?.id}
-            disabled={!!selectedProgram}
-            onChange={({ selected }) => setSelectedProgramId(selected)}
-            placeholder={i18n.t('Choose a program')}
-            maxHeight="max(60vh, 460px)"
-            dataTest={'program-select'}
-        >
-            {programs.map(({ id, name }) => (
-                <SingleSelectOption key={id} label={name} value={id} />
-            ))}
-        </SingleSelect>
-    )
 
     return (
         <div className={styles.rows}>
             <div className={styles.columns}>
                 <div className={styles.stretch}>
-                    {!selectedProgram ? (
-                        select
-                    ) : (
-                        <Tooltip
-                            content={i18n.t(
-                                'Clear program first to choose another'
-                            )}
-                        >
-                            {select}
-                        </Tooltip>
-                    )}
-                </div>
-                {selectedProgram && (
-                    <Button
-                        small
-                        secondary
-                        onClick={() => setSelectedProgramId(undefined)}
-                        dataTest={'program-clear-button'}
+                    <SingleSelect
+                        dense
+                        selected={selectedProgram?.id}
+                        onChange={({ selected }) =>
+                            setSelectedProgramId(selected)
+                        }
+                        placeholder={i18n.t('Choose a program')}
+                        maxHeight="max(60vh, 460px)"
+                        dataTest={'program-select'}
+                        filterable
                     >
-                        {i18n.t('Clear')}
-                    </Button>
-                )}
+                        {programs.map(({ id, name }) => (
+                            <SingleSelectOption
+                                key={id}
+                                label={name}
+                                value={id}
+                            />
+                        ))}
+                    </SingleSelect>
+                </div>
             </div>
             {showStageSelect && (
                 <StageSelect stages={selectedProgram.programStages} />
